Run build and tests with multiple JDK releases
