+++ lib/src/test/java/fi/protonode/certy/TestCredentialJdk8.java
@@ -49,7 +49,6 @@ import java.security.cert.CertificateException;
 import java.security.cert.CertificateFactory;
 import java.security.cert.X509Certificate;
 import java.security.interfaces.ECPublicKey;
-import java.security.interfaces.EdECPublicKey;
 import java.security.interfaces.RSAPublicKey;
 import java.security.spec.ECParameterSpec;
 import java.security.spec.InvalidKeySpecException;
@@ -60,7 +59,10 @@ import java.time.temporal.ChronoUnit;
 import java.util.Arrays;
 import java.util.Date;
 
-class TestCredential {
+/**
+ * Test cases for Credential class using JDK 1.8 and later.
+ */
+class TestCredentialJdk8 {
 
     @Test
     void testSubjectName() throws Exception {
@@ -113,16 +115,6 @@ class TestCredential {
         expectKey(cred.getX509Certificate(), "RSA", 2048);
         cred.keySize(4096).generate();
         expectKey(cred.getX509Certificate(), "RSA", 4096);
-    }
-
-    @Test
-    void testEd25519Certificate() throws Exception {
-        Credential cred = new Credential().subject("CN=joe")
-                .keyType(KeyType.ED25519);
-        X509Certificate cert = cred.getX509Certificate();
-        assertNotNull(cert);
-        EdECPublicKey key = (EdECPublicKey) cert.getPublicKey();
-        assertEquals("Ed25519", key.getAlgorithm());
     }
 
     @Test
@@ -266,9 +258,6 @@ class TestCredential {
 
         Credential cred2 = new Credential().subject("CN=joe").keyType(KeyType.RSA).keySize(1);
         assertThrows(IllegalArgumentException.class, () -> cred2.getX509Certificate());
-
-        Credential cred3 = new Credential().subject("CN=joe").keyType(KeyType.ED25519).keySize(1);
-        assertThrows(IllegalArgumentException.class, () -> cred3.getX509Certificate());
     }
 
     @Test
@@ -423,10 +412,8 @@ class TestCredential {
         assertNotNull(dps);
 
         DistributionPointName expected = new DistributionPointName(
-            new GeneralNames(
-                new GeneralName(GeneralName.uniformResourceIdentifier, "http://example.com/crl.pem")
-            )
-        );
+                new GeneralNames(
+                        new GeneralName(GeneralName.uniformResourceIdentifier, "http://example.com/crl.pem")));
 
         assertArrayEquals(new DistributionPoint[] { new DistributionPoint(expected, null, null) }, dps);
     }
